Fix saving user to Firebase
