+++ data/src/main/java/com/example/data/repository/FirebaseAuthRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.example.data.repository
 
 import com.example.data.mapper.UserModelMapper.mapToUser
+import com.example.data.mapper.UserModelMapper.userToMap
 import com.example.domain.model.User
 import com.example.domain.repository.FirebaseAuthRepository
 import com.google.firebase.auth.FirebaseAuth
@@ -18,61 +19,60 @@ class FirebaseAuthRepositoryImpl @Inject constructor(
         private const val USERS_KEY = "users"
     }
 
+    private val usersCollection = firestore.collection(USERS_KEY)
+
     override suspend fun signInWithGoogle(idToken: String): Result<User> {
-        val credential = GoogleAuthProvider.getCredential(idToken, null)
-        val authResult = auth.signInWithCredential(credential).await()
+        return try {
+            val credential = GoogleAuthProvider.getCredential(idToken, null)
+            val authResult = auth.signInWithCredential(credential).await()
 
-        val firebaseUser = authResult.user ?: return Result.failure(Exception("User is null"))
+            val firebaseUser = authResult.user ?: return Result.failure(Exception("User is null"))
 
-        val isNewUser = authResult.additionalUserInfo?.isNewUser ?: false
-        val userId = firebaseUser.uid
+            val isNewUser = authResult.additionalUserInfo?.isNewUser ?: false
+            val userId = firebaseUser.uid
 
-        val userDocument = usersCollection.document(userId)
-        val targetCalories: Int
+            val userDocument = usersCollection.document(userId)
 
-        if (isNewUser) {
-            targetCalories = 0
-            val newUser = User(
-                id = userId,
-                name = firebaseUser.displayName,
-                email = firebaseUser.email,
-                photoUrl = firebaseUser.photoUrl?.toString(),
-                isNew = true,
-                targetCalories = targetCalories
-            )
-            try {
-                userDocument.set(newUser).await()
-            } catch (e: Exception) {
-                return Result.failure(e)
-            }
-        } else {
-            try {
+            if (isNewUser) {
+                val newUser = User(
+                    id = userId,
+                    name = firebaseUser.displayName,
+                    email = firebaseUser.email,
+                    photoUrl = firebaseUser.photoUrl?.toString(),
+                    isNew = true
+                )
+                val createResult = createUser(newUser)
+                if (createResult.isFailure)
+                    return Result.failure(createResult.exceptionOrNull() ?: Exception("Failed to create user"))
+
+                Result.success(newUser)
+            } else {
                 val snapshot = userDocument.get().await()
-                val existingUser = snapshot.data?.let { mapToUser(it, userId) }
 
-                targetCalories = existingUser?.targetCalories ?: 0
-            } catch (e: Exception) {
-                return Result.failure(e)
+                val existingUser = snapshot.data?.let { mapToUser(it, userId) }
+                    ?: return Result.failure(Exception("User data not found"))
+
+                Result.success(existingUser)
             }
         }
+        catch (e: Exception) {
+            Result.failure(e)
+        }
+    }
 
-        val user = User(
-            id = userId,
-            name = firebaseUser.displayName,
-            email = firebaseUser.email,
-            photoUrl = firebaseUser.photoUrl?.toString(),
-            isNew = isNewUser,
-            targetCalories = targetCalories
-        )
-
-        return Result.success(user)
+    override suspend fun createUser(user: User): Result<Unit> {
+        return try {
+            val userMap = userToMap(user)
+            usersCollection.document(user.id).set(userMap).await()
+            Result.success(Unit)
+        } catch (e: Exception) {
+            Result.failure(e)
+        }
     }
 
     override suspend fun signOut() {
         auth.signOut()
     }
-
-    private val usersCollection = firestore.collection(USERS_KEY)
 
     override suspend fun isUserLoggedIn(): Boolean {
         return auth.currentUser != null
