+++ domain/src/main/java/com/example/domain/repository/FirebaseAuthRepository.kt
@@ -6,6 +6,7 @@ interface FirebaseAuthRepository {
     suspend fun signInWithGoogle(idToken: String): Result<User>
     suspend fun signOut()
 
+    suspend fun createUser(user: User): Result<Unit>
     suspend fun isUserLoggedIn(): Boolean
     suspend fun getCurrentUserId(): String?
     suspend fun isUserFullyRegistered(userId: String): Boolean
